refactor: use new scoped function
